requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic
openai>=1.12.0
python-dotenv==1.0.1

langchain-openai==0.1.7
langchain-community==0.2.0
langchain-core
faiss-cpu==1.13.1
tiktoken
